Fixed data loading problem in TreeGrid
Due to the async nature of Tabulator creation and certain APIs like setData it can happen that quick/early updates of the TreeGrid component don't reach the underlying tabulator component. The previous mix of controlled and uncontrolled component behavior made things worse.

This has been reworked to make TreeGrid fully uncontrolled (data is held in tabulator cells). New data is only held for a short period of time if the tabulator initialization is still in progress when the data comes in. Additionally, it's still possible to provide columns + rows as properties, for the initial creation of a TreeGrid instance (useful only for static content that never changes, like settings lists, dialog tables and so on). This data cannot be updated.

A number of other changes in this patch:
- Quicker result view updates when the height of the result area is already known (for example when restoring previous results).
- It's no longer needed to have different handling for the table member of TreeGrid, depending on the result (tabulator instance vs promise). It's now always a promise.
- Clean up of result set + text result data interfaces for script results.
- Fixed some tests.

Change-Id: I26afb11178154a39fda49000ef0099cbd26c675e