Workaround for a display bug in shell sessions
This is a workaround for JS and PY modes in shell sessions, which do not return the required column info (as it happens for SQL mode). So for now we get column names from the first data row (if any) and assume string for it.

Also: dispose of the Monaco instance when a CodeEditor is unmounted.
Change-Id: Ie6ad4cd31bae1231eb4aa7ee7b1fff0c741b2a34