This patch is to test the screen resolution problem
The problem is that we use psexec to trigger the tests on a session which has access to the UI, which does not provide a way of changing the screen resolution (which is too low). This forced us to use headless mode. Since the extension tests do not use a browser, headless mode is not available. So we try to increase the screen resolution of Session 1 using powershell

Change-Id: Id5639b3ec4c3e31349a7031af88a0034b35702c0