Work with docker
This patch refactors the code to work with the remote server: http://docker.mysql.oraclecorp.com/
With this patch, we no longer stop/start the shell server for each test on some test-suites (previously, it was used to have a clean environment at each test). Also, the powershell script no longer has to download Shell UI and setup the whole environment. At the end, we expect a better performance when executing tests.

Change-Id: Ic0ac93996ce71b93767417c356fd485b0828c993