Removed reconnecting-websocket module
For Node.js (vscode extension use) we have to specify a root CA certificate to validate our self signed certificate used by the shell. This is especially useful to simplify the cert setup on Linux.
Additionally, the reconnecting-websocket behavior did not meet our required connection handling.

Because of these reasons it has been removed and the previous setup has been restore. Later we have to take care ourselves for automatic reconnection and other problems related to that.

Change-Id: Ida8196e9b6d80850838ab68cfc2d007b3cd355c8