Added option to cancel dialogs
Cancelling dialogs was possible before but was considered the same as the "no" answer in dialogs that have a yes/no semantic. This has now changed to allow the "no" answer to trigger something else, instead of meaning to cancel a process.

Additionally, there's now a close button in dialogs.

Change-Id: If000c8966c360ae468b607920a671e10c326e8d4