Bug #34271996 About inserting Japanese data
When attempting to insert japanese characters using the UI the inserted data looked
like some sort of binary data, querying the inserted data showed the corrrupt information.

The problem was caused by a wrong encoding of the data received by the WebSocket in the
backend, which was converting each received byte to characters without considering the fact
that some UTF-8 strings contain multibyte characters.

This patch fixes the problem by properly encoding the messages to UTF-8, which leads to the
correct interpretation of multibyte characters.

Change-Id: I94722ad9ab8500aa3e3a6e6b5b00872dba344eaf