Reworked temporary solution for unique column identifiers
Previously fields in result data were based on the column's title. However, depending on the query there can be duplicates (e.g. `select 1 as a, 2 as a, 3 as b, 4 as b, 5 as a from DUAL`, hence a solution was implemented to add running numbers to column titles. However, this doesn't always work depending on what's actually in a column title. Instead the handling has been changed to use the ability of the TreeGrid to use a field identifier to find a column's value in a row, which is separated from a column's title.

Change-Id: I010d4d0eda38d3e7904afee1ac061d049259def2