Misc bug fixes including Bug#34504281 Bug#34241454
Bug #34504281 Shell Console: Error messages on connecting to MySQL Server at localhost

Even the user session was correctly established, the service session was being created
using incorrect connection data, now the correct connection data is used.

Bug #34241454 SCHEMA LIST IS NOT DISPLAYED ON SHELL SESSION

The problem is caused because the schema list is retrieved using the service session
which is created based on the prompt descriptor that comes from the shell which does
not include the password. When the password is not included on the connection string
a prompt will be used to resolve it, giving the opportunity to store it for use
when creating the service session. The problem was solved on this very specific case
by getting the password from the executed command.

Additional Fixes:

- Updated bastion name generation to be based on timestamp.
- Hidden password prompt replies from debug log on Shell Console.

Change-Id: I7748ddc5d9bcf8ef898df85c14c80f9bc15de43b