Don't wait for the execution of commands in DB editor
This patch changes the behavior in the DB editor such that it no longer waits for long running tasks before it creates the next execution block. However, it's still necessary to finish the running task before a new one can be started.

Also the command to update jest snapshots has been fixed (was missing the test setup).

Change-Id: Ib38d224b8e4a8097e958fe18fa6461c207590ad0