Fixes usage of sqlite cursor accross threads.
The introduction of the ResponseHandler to process FE requests moved the execution
of the functions to a separate thread.

Even SQLite allows sharing connections across threads, the Python Sqlite cursor
requires proper control to avoid race conditions on the usage of cursors.

This patch fixes the problem by adding a thread lock before the execution of
statements on an SQLite connection.

Change-Id: If1bcc2ff08c0f876947bdb9c4bce5d8ef26dcdf2