Fix to disable DBPingHandler while a task is being executed on a Session
Long running SQL Tasks were being interrupted by the DBPingHandler,
since it runs on a separate thread, it was triggering the execution of
the ping SQL while a different query was being processed using the same
session.

This was specifically problematic when the session was being used by
external (non GUI) plugins, since they use the inner session directly
rather than scheduling tasks in the Queue.

This patch fixes this problem by having the DBPingHandler register a
callback so the Session can notify when a task starts and ends, this way
the DBPingHandler is PAUSED while the task is executed.

Additionally, logic was added to effectively disable the DBPingHandler
once the Session is closed.

Change-Id: I877237b6a8533f652364f4cbd771c2334f03e065