E2e - Start inactive dbsystem and skip test
We have been facing an issue with the MDSforVSCodeExtension DBSystem, which is sometimes stopped (I don't know why).
This causes a test to fail, giving us the idea of something wrong. Since OCI is an external system, it does not make sense to fail the test because of an inactive DBSystem.
This patch adds the following logic:
- If the DBSystem is stopped, start the DBSystem and skip the test. We don't want to wait too much time for the start, which usually takes a bit. This is allow that the test can be executed on the next patch run.
- If the DB System is active, proceed with the test

Change-Id: I675a68f7177a137cbc07ddcaa291bcfea8d62b5f