Fixed bug for prompts during opening a session
While a session (shell or DB editor) is being opened no module session id is known yet (it's returned from that opening call), but there can be prompt requests (e.g. to enter a password). These prompts carry a module session id, which can be used to send replies to the backend.

Change-Id: If3f857a46acafa61ef0e6f4c1eb7fba88af1ba9c