Separate connection browser and connection editor
The connection editor started as a few lines in the connection browser, as it was only used there. Over time the editor grew beyond all expectations, with more than twice the size of the browser, so it was time to extract it into an own class.

This patch also removes the Node.js "path" import, which was only necessary for a simple task (which has been implemented as an own helper method).

Change-Id: Idd013bda8d437f2ed120d2505a2b0dc900024221