WL#15638 Open Editors View in VS Code Sidebar
VS Code's Explorer section maintains a view "OPEN EDITORS", which lists all currently opened editors for quick navigation.

This Worklog implements such a view for the MySQL Shell for VS Code extension. By default this new view is placed a the top of the side bar. With its inauguration the "MYSQL SHELL CONSOLES" view becomes obsolete and is removed.

Change-Id: I767809f814c1d82d6d6d87234e28c537140b04fd