Bug #35115512 Wrong default value for Rest object pagination in MRS
There were a couple issues regarding this patch. It was not possible to set the items_per_page to null and, on the other hand, adding a db object was setting items_per_page to 25 in the absence of a value.

The solution is to set the items_per_page as a required parameter to allow the frontend to set it to null. Also, the backend can not set it when the value is empty or zero.

Another issue that came up was that setting CRUD operations was mandatory on update. This doesn't seem to be necessary, so it's only updated when supplied on the update values.

Specific tests were also made for this specific situation.

Change-Id: Iaa5d7bf138f5bda8eb1fd2335057588b01b058c8