Fix WelcomeWebview TypeScript compilation error
When checking if a given platform version is supported, the numeric
version is extracted using the "os.release()" API via a regular
expression, and even though it should always return a valid match, there
is a conditional that checks if the match array length is greater than
1, which potentially could lead to a situation where a boolean value is
returned if that is not the case.

The TypeScript compiler correctly identifies this scenario and yields a
compilation error.

Change-Id: I82a7f2b7eba1d2db90d0ea4dbf07965786884db6