Fixing race condition.
This patch fixes a specific race condition in which a MySQL Session
could be closed by one thread (i.e. when the Web Session is terminated)
while a statement is being executed in a separate thread. Since the
sessions are not thread safe in the Shell, this condition leads to a
unexpected termination of the Shell running the backend server.

Change-Id: Iaf091fd4e7c62abe87b5a60a75042b493bab33d1