BUG#35124859 BUG#35124804 BUG#35124746 Options marked as required make the plugins unusable in the MySQL Shell
Many options in several plugin functions (GUI/MRS Plugins) have been marked as 'required' as a way to generate the options as nullable in the
TypeScript protocol files. Even this is a good approach for the TypeScript interfaces, it causes the plugins to be unusable in the Shell as it causes
options that should be optional, to be mandatory. This patch removes the 'required' and 'optional' flags from the plugin function documentation
and moves the definition of nullable options to the protocol generation script.

Change-Id: I55d92a63f3144f68c68716d12cac532b53e8e995