BUG#35317260 Default value for url_protocol when adding REST service
In order to provide a better developer experience when adding a new REST
service using the MySQL Shell in batch mode, the "url_protocol"
parameter, which is not required, needs to be assigned a sane default
value.

In interactive mode, it should also be possible to override the default
behaviour by providing an empty set of options.

This patch introduces the changes to accommodate these expectations.

Change-Id: Ie0a0da49750a1f03e76a65a02637e760e3873516