WL#15683 Implement enhanced Notebook support
Currently result sets and their visual representation as tables or graphs are stored only for the current session. The editor used for them is called a notebook and provides an extraordinary user experience with its graphical result views and charts.

This worklog is about creating an even better experience by making notebooks own entities, which can be made persistent, instead of only being ad hoc documents. This will allow to:

- Suspend and resume ongoing work in complex scenarios.
- Keeping important sets for later reuse in demos or teaching in classes.
- Share sets in a team for review + discussion.
- Organize notebooks in a larger project.
- Restore a session where the user left.
- and more...

The key for this extended handling is the introduction of a new file type (*.mysql-notebook), which is a special JSON document with just enough data to restore the notebook, including any loaded result data and last editor state (selection, caret position and more). Additionally, the extension implements an own custom editor for these files, which allows synchronization between a notebook file and a notebook editor.

Change-Id: Ibee51fd1b3a21f7fd829ce9e402e46b3cb153149