BUG#35377927 Delete REST schema after proper confirmation
A confirmation dialog is displayed to users when they try to delete an
existing REST schema using the MySQL Shell GUI Extension for VS Code.
The schema should only be deleted if the answer is "Yes".

Currently, the schema is being deleted regardless of the answer to the
confirmation dialog. This patch addresses that issue and ensures the
schema is only deleted when the answer is "Yes".

Change-Id: I57dfa9fe6ab0265a62fbb592c1cc8d8186777033