BUG#35381777 Fix service metadata display format
In an interactive MySQL Shell session, the
"mrs.get.current_service_metadata()" command should always display the
metadata details using a tabular format whereas in when the MySQL Shell
is used to execute a script in batch mode, the command should return a
JSON-like structure. Currently, the command returns that JSON-like
structure regardless of the MySQL Shell session type.

This patch introduces the changes to ensure the
"mrs.get.current_service_metadata()" command displays the details using
a tabular format when called in an interactive MySQL Shell session.

Change-Id: I28665c4e58418d282b35c0da97728728e36be3ca