BUG#35381360 Display service listing for "mrs.get.service()"
In an interactive MySQL Shell session, the "mrs.get.service()" command
should always display the service details using a listing format. If the
command is called without any argument, the user is prompted to select
the service from a list. After properly selecting the service, the
command returns back a JSON-like structure containing the service
details.

This is the expected behaviour when using the MySQL Shell to execute a
script in batch mode, but not in interactive mode. This patch introduces
the changes to ensure the "mrs.get.service()" command displays the
service details using a listing format when the service is selected from
the prompt in interactive mode.

Change-Id: I044161156da9513b6ad4b27ee4d63e71c815f3f3