Use Symbolink Links on the test suites
Currently, when running the ext tests in parallel, we need to install the extension 4 times, on each VSCode instance. Although this task is made in parallel, to simplify the process, we now have to install the extension once, on one VSCode instance, and then use symbolic links the that extension on the other VSCode instances. It means that only only Shell process will be used to serve the tests, instead of 4.

Change-Id: Ieca54c4df48053f3e71dc5a092fbec32c01cab84