Bug fix for wrong shell config dir
This patch contains a better implementation to determine the shell config user dir, based on how the extension is executed. Previously it was based on where the mysqlsh binary was found, which can greatly differ among platforms. This patch now uses a flag in the extension context to figure out if we are debugging (and have to use the standard shell config dir) or not (and use the dedicated .mysqlsh-gui folder).

The config dir can still be overriden by either specifying a path in env.MYSQLSH_GUI_CUSTOM_CONFIG_DIR or as an app parameter (for testing).

Change-Id: I8f6c3acda36237375532fc470d3fd86f86b93b01