BUG#35594274 Fix unexpected error with orderBy
The MRS SDK "find*()" API should allow to order (ascending or
descending) the records in the result set based on a given field using
the "orderBy" option. However, using this option without a specifying a
filter with the "where" option yields an unexpected error.

This patch fixes the issue and introduces a couple of unit tests to
prevent future regressions.

Change-Id: Ic8bb2f1b158827d3ac69c5e620a3ca756b1ba2e0