BUG#35594404 Remove extraneous "findFirst()" options
The MRS SDK "findFirst()" API is used to retrieve the first record that
optionally matches a given filter, and additionally, select specific
fields from that record. Currently the API supports extraneous options
like "take" and "fetchAll" that only make sense when operating on
multiple records.

This patch removes the extranous options from the "findFirst()" API and
introduces automated tests for the corresponding types, as well as for
the behaviour of the API itself and its different SDK counterparts like
"findMany()" and "findUnique".

Change-Id: I6c38b9b780377e1c275859d19e21f7e817ebba34