BUG#35635932: Enable type scaffolding for "find*" options
The "findFirst()", "findMany()" and "findUnique()" commands in the MRS
SDK did not support auto-completion and type-safety for their
corresponding input options. This was because the template still
referenced older type definitions, which have been renamed recently.

This patch updates the template to import the correct type definitions
and only when they are effectively needed, based on the types of CRUD
operations that are enabled for the database objects available in the
REST service.

Additionally, the new template has been used to update the SDK files in
the MRS Notes sample project.

Change-Id: I9ace938339d9722dd5b27c865c333052e9739538