Fixing BE Tests - Avoid creating webroot within the code folders.
The script that executes the BE tests was creating a remporary webroot directory
in the source code to satisfy the requirement of the webroot folder for the test
server. This patch removes that logic and updates the test_server.py file to
use an existing webroot if found on the source path, or create a temporary webroot
in TMP otherwise.

Change-Id: I2427a4989d20e69988e3d0e087833e19ec713171