BUG#35645706 Add proper mapping for JSON datatype
Whilst generating the MRS TypeScript SDK, columns with the "JSON" data
type are being mapped to "unknown". This results in compilation errors
due to potential excessively deep and possibly infinite type
instantiations.

This patch introduces a new type definition for JSON values that is now
being used to map existing JSON columns in a database object.

Change-Id: I2b5486d92482aa2bbc3d0e89b4447bf8f573907a