WL#15880 Handle "NOT NULL" constraint in the MRS SDK
MySQL supports "NOT NULL" constraints which ensure that the value in a
given column cannot be "NULL". By omission though, a column can hold
"NULL" values. With the MySQL REST Service, records containing columns
with "NULL" values can be included in or excluded from the result set
using the "$null" or "$notnull" operators.

This patch introduces the changes required to ensure the TypeScript SDK
provides a consistent API (with auto-completion and type-safety) to
filter for "NULL" values. Additionally it also introduces the
constraints required to yield a proper compilation error when a user
performs a "NULL" value lookup for a field which maps to a column with a
"NOT NULL" constraint.

Change-Id: I89cb106817da43b39eebc215c74dba1a1083e71e