Fix Extension tests
This patch fixes the issue with the extension e2e tests. There is an opened issue on the selenium webdriver (https://github.com/SeleniumHQ/selenium/issues/12481) and this patch installs a previous version of it, to workaround the problem until it is fixed. The webdriver is a dependency of the vscode-extension-tester

Change-Id: I51be5019afa5bd2baf57b21b8b5b5ba49e919557