WL#15912 Add MRS SDK types for spatial columns
Spatial column data types are currently mapped to a generic MRS SDK
client side type in the form of a JSON object. Although this works,
because it still allows to represent data using the GeoJSON format, the
type definition itself is not helpful from a standpoint of
auto-completion hints and does not account for other popular standard
spatial types such as Well-Known Text (WKT) which is an official OpenGIS
representation.

This patch introduces formal GeoJSON and WKT type definitions in the MRS
SDK for each spatial data type supported by MySQL.

Change-Id: I91bc708458ed3f248042be37a24902d604a50e42