BUG#35745272 Add client-side type for spatial columns
The MySQL router only understands spatial columns specified in GeoJSON.
From the MRS client standpoint, the most appropriate mapping data type
for such columns, in TypeScript, is a plain object.

Currently, the MRS TypeScript SDK is mapping the GEOMETRY column data
type to a type definition that represents an object with "unknown" field
values. This leads to potential compilation errors due to excessively
deep and possibly infinite type instantiations. Additionally, other
spatial data types are being mapped to a string, which needs to be
parsed, and unless it contains valid JSON, it is incompatible with the
format supported by the MySQL Router.

This patch introduces the changes to ensure the MRS TypeScript SDK
starts mapping any spatial data type into a proper JSON object
representation on the client side.

Change-Id: Ieee339d5efc41aa08ff3e9027831ca61b48759b0