Exchanging the ANTLR TS runtime
Because the previous runtime (antlr4ts) is no longer updated and is missing some features from the latest ANTLR4 release, no new 3rd party approval is possible. So we want to and have to use a new ANTLR4 runtime. This requires a number of changes, both in the extension as well as the frontend.
This required also to update the antlr4-c3 package.
Additionally, some unit tests have been fixed, the MySQL grammar was simplified (some direct left recursion and ambiguities removed) and the Python lexer base class has been completed. The MySQL parse service test suite includes now a warm parser run.

Change-Id: I87b84d4c26631f9ec749f32da7797292d0c80acb