Extension e2e OCI configs
This patch points the OCI config file of all the test-suites different than OCI to a dummy config file. This will prevent the extension to read the existing data of oci configs on the default folder location, which contains information for all jenkins agents. This was obligating the extension to always read the OCI configs for other test-suites that don't need it

Change-Id: I96c119116771af688c1cf93ebbd8837fbf58f74b