BUG#35848217 VSCode Extension does not start with Python 3.14
The database access code in the gui backend plugin is implemented using
an interface that mimics the database interface defined for the MySQL
Session objects in all products.

To achieve this in SQLite database access, custom versions of Connection
(SqliteConnection) and Cursor (DbCursor) are used as class factories which
is functionality supported by the SQLite Python API.

With python 3.11.4, the usage of those facories got broken, in a way
that the backend code produces sqlite3.Cursor instances instead of the
custom DbCursor instances, leading to several errors when accessing
SQLite databases.

This patch fixes the problem by ensuring the custom classes are used
instead of the default ones provided by sqlite3.

Change-Id: I23e28b72873bb298c25cda45a2b02b8c7d780148