Fix bad and failing MRS SDK type definition tests
The test for GeometryCollection specified as a GeoJSON object is failing
due to the way "@ts-expect-error" works. Since it applies only for the
preceding line of code, the type-checker will happily validate
everything that comes after it.

Additionally, some tests are incorrectly using "toEqualTypeOf()", and
not really executing the assertion, which in fact, would also be
failing, because it is not the intended check. What those tests want to
really validate is if the type also matches a different definition.

In this case, due to a linter restriction on the size of each line of
code, an object had to be spread with fields specified on multiple
lines. The test was expecting a type check error for only one of the
fields since the other field, specified on a subsequent line  was
correctly specified. In this case, the solution can be to use
"@ts-ignore" for that valid field, in order to ensure the test works as
expected.

This patch fixes the failing test and replaces the incorrect instances
of "toEqualTypeOf()" by the more appropriate call to "toMatchTypeOf()".

Change-Id: I909db6a8be00c236e937c3a7de05e8f4dba49d1b