Combined MRS + MySQL grammars for use in the FE
This patch contains a new parser + lexer which is a combination of the simple MRS grammar from the MRS plugin and the standard MySQL grammar, to enable MRS syntax in the frontend. The MRS syntax is not an official MySQL language addition, but supported by the MySQL shell.

This also involves adding a number of new keywords and a flag to switch MRS support on and off.

Change-Id: I6663a25ff3a878ad4af5a43c37a66511063e5b8c