Converting MSG to ESM only
This patch is a preparation for cases where we have to import code directly into webviews and it is generally a good idea to fully advance to ESM. VS Code (Node.js) still uses CommonJS, but that's handled by our building system (based on esbuild). ESM support in Node.js is already working and just a matter of time to become the main module type everywhere.

But because browsers cannot automatically resolve imports these have to be explicit, which means: the exact path plus the script extension (usually .js). With TypeScript a new moduleResolution option was introduced recently (`Node16`) which enforces this restriction. Hence most of the changes in this patch are adding the required extension for this module resolution type.

I also moved a preact helper method out from the TS helpers to avoid importing Monaco and TS when using the preact stuff.

Change-Id: I4331f6df8cc8a052133850b5de9bbfcb0c5d1262