Extending the way code is executed
Previously we only executed SQL code in execution context. With this patch we can run code in all supported languages. It also builds on the recently added code broadcast feature and refines it. With that in place we can also clean up duplicate interfaces.

Additionally, a new method on the SQLExecutionContext class allows to set new text on the underlying editor model and wait for the context until statement splitting is finished.

Change-Id: I19fcaf2b5551361cb90d222e38d488f2d9437d96