Fix typecheck errors in the MRS SDK files
This patch fixes some newly-encountered typecheck errors in the MRS SDK
files.

The SDK can now be generated using a Preact-based blueprint and  thus it
requires the package to be available, at least as a devDependency to
avoid typecheck errors about missing dependencies, something which is
also formalized by this patch.

Change-Id: I8158f9aa8a642f79d933a3bd809efaff8774f91a