BUG#36032142 Add support for nullable filters in MRS SDK delete*()
Deleting records with the MRS TypeScript SDK does not work when a filter
containing NULL checks is used. This is because the translation of the
underlying operator in the REST API is not being done correctly.

This patch introduces the changes required to introduce support for NULL
checks in query filters used by DELETE operations.

Change-Id: Iff535441e57cd13d5f77a5e2f873afecf04d748d