Remove extraneous dependencies for MRS SDK types
The TypeScript MRS SDK types mapped from column data types should be
generated as is, based on the field information extracted from the
database object.

Currently, the type declarations are mixing in extraneous details
related to REST or ORDS specific fields, which do not apply for user
input when the data flows through the SDK.

If an object needs to be decorated with REST or ORDS specific fields,
that should happen when the response is being built.

This patch removes the extraneous type dependencies and moves the type
declaration and inference logic to the proper place. Additionally, it
updates the sample project to use the most recent version of the SDK.

Change-Id: I0e87f0fdef28891e98a186bb45339188b7242dca