Fixing shell executable detection.
Newer versions of the  MySQL Shell are now saving the binary path
in mysqlsh.executable since sys.executable points to the python
binary. This change was done so pip can properly handle the
installation of any module.

This patch updates the logic to properly handle the shell binary
identification in any shell version.

Change-Id: I085e050de55306e3c98c32cea04c3cad1df5a29b