Remove usage of OCI environment variables
This patch removes the usage of so many environment variables to set the oci profile configuration. Instead, we now only use one (MYSQLSH_OCI_CONFIG_FILE), which points to the location of the configuration file. This file should be fulfilled by the user.
Other enhancements were made on the OCI tests code, to fix the acceptance of a generic OCI profile to be used on the tests

Change-Id: I1c06e16faee710e8e9b8f8cdffe2fab028726f2d