Full handling of clipboard operations in embedded scenarios
Clipboard handling in an embedded iframe is a bit tricky, as the iframe cannot directly access the clipboard and the app host cannot just forward clipboard events to the iframe (for security reasons).
That's why we have to handle cut and copy manually and send the text for the clipboard via requisition to the host, and handle inserting text on a paste event, which is sent by the host.

Note: this is only for input elements from our UI framework and standard edits. For the Monaco editor separate handling is necessary, which mostly consists of registering keyboard hotkeys (which was implemented in a different patch).
Change-Id: I54e8560825d45a807a339a6f2d4fc49d65df6813