Add user friendly errors messages to Extension tests - 1st batch
This patch is the first batch of new error messages to expectations. These error messages are more user friendly. Before, we could see something like "expect true to be false" on failed expectations, not knowing exactly what was wrong when reading the test report. Now we have clear messages when something fails

Change-Id: Iac6b06f44abd7f45c18da22be8023eee64fdd177