BUG#36173373 Fix function/procedure "fetch()" declaration
In TypeScript, properties/fields in a class such as arrow functions
become part of "this" which makes it impossible to call "super" when
overriding those functions in a potential child class.

The most recent versions of the TypeScript compiler actually yield an
error when a method in a parent class is declared with an arrow function
and then is called by a method in the base class.

This patch updates the "fetch()" method declaration in the classes
responsible for handling function and procedure calls (as well as the
base class definition) to avoid bumping into this compiler error.

Change-Id: Ib0f68e8f56dd206bc55ae26531a11914a317bc63