Introducing minimum and expected shell versions
The FE now defines two version numbers:
- The minimum number which is needed for the app to work.
- The preferred version, which is the version the app prefers during testing.

With this solution we can change the shell on test hosts without having to change the version test (unless we want that). And while we are at version checks it makes sense for the app to check if the shell it connected to is what it needs.

Another fix in this patch is the removal of old react eslint settings, which made the linter stop working.

Change-Id: I1990bf6744519b34d0e762fc090c8070405a7fb8