BUG#36211402 Avoid field assignment on "super"
TypeScript/JavaScript uses prototype-based inheritance where fields are
stored in a flat structure and not in the scope of a particular class
definition in the hierarchy. This means that a field specified by a
parent class can be replaced by one specified by one of the children.

To signal this behaviour, starting in v5.3.0, the TypeScript compiler
prevents a class from assigning values to fields specified by a parent
class using "super"
(https://github.com/microsoft/TypeScript/pull/54056).

This patch, upgrades the TypeScript compiler to the latest version and
fixes the remaining issues related to the topic mentioned above that
now result in compiler errors. Additionally, it also updates the launch
task that runs the compiler to ensure it uses a local version of the
compiler, as specified by the dependency graph of the VSCode Extension
package.

Change-Id: Ib3ef774028ec4587c6b8e5eaa2eb395188e0d5a1