BUG#36204648 Fix OR syntax with existing implicit AND
The MRS TypeScript SDK provides a fluent API for constructing HTTP
requests containing query filters with a syntax supported by the MySQL
Router. The API includes a utility function "or()" which should be
capable of extracting the existing filter context into a high-order OR
operation.

If the query filter only contains top-level implicit checks of the type
"field: <value>", the MySQL router returns an unexpected SQL error for
the generated request. This is because, the router still has some
limitations with regards to the various syntax formats supported on
ORDS.

This patch updates the query filter generation logic to ensure the
request only uses syntax supported by the MySQL Router.

Change-Id: Ie6041a77d4bddb0ff98bc7cf7894045a1da36def