Fix column metadata 'default' value bug
Because there is no way to distinguih between `dafault` set to null
versus skipped this patch changes the returned metadata values
and not includes 'default' key if `not_null` is true and `default` is null.
In addition to that, unit test have been fixed and new ones have been added to cover
more situations with different values 'not_null' and 'default'.

Change-Id: I021465f3d9cf631281af973cf1a86cb107ac57c9