Remove OCI test
This patch removes an OCI test - Edit an existing MDS Connection. Some time ago, we had 2 DBSystems available for our test suite, but now, due to resources costs, we only have 1. This test, edited an MDS connection to change a connection DB system, but now, since we only have 1 DB system, the test fails, and it has no sense to keep having it on the test suite.

Change-Id: I9dd5323c48727b1e8115632f56eb0caff49ffbe8