Workaround for vscode-extension-tester arm64 downloads on macOS
vscode-extension tester has an issue: When the system has a arm64 CPU, it downloads vscode for non-arm64 cpu. This results on the tests being very slow and unstable.
This patch includes a workaround to overcome this situation - If the system is an arm64 macos system, we perform the download manually for the correct cpu arch

Change-Id: I27fbb8b20d20fb9dc8897665ac92a123db17d288