BUG#36347569 Fix help links in the MSG front page
Clicking the links available in the front page of the MySQL Shell GUI
displays a dialog to add a new REST object instead of opening a new
browser window/tab in the URL encoded by the corresponding hypertext
reference. This is caused by a custom debug event listener that
explicitly handles click events on those links and forces the dialog to
be displayed instead.

This patch removes the extraneous custom debug event listener and
additional dependencies.

Change-Id: I48bb4cc21cb9f19a1b43d16da7406d5806e75a2b