WL#16334: Add support to generate base Python SDK
The MySQL REST Service (MRS) Software Development Kit (SDK) generated for
a given REST service includes a set of constructs and APIs that simplify
workflows for retrieving from and sending data to the resources available
as part of that REST service.

A big goal of the SDK is to provide end-to-end types for database object
fields, ensuring the client is always aware of the MySQL data types used
to represent each item in the database. The type-safety infrastructure
unlocks additional auto-completion and validation features which are
usually available in Language Server Protocols (LSPs) and offer a better
developer experience for anyone writing MRS-based applications.

Python is a very popular language which allows to write this kind of
infrastructure leveraging a type checker like MyPy. Applications written
in Python on top of MRS should also benefit from this level of developer
experience.

This patch bootstraps the Python MRS SDK, with an API inspired by Prisma
(similar to the MRS TypeScript SDK). Developers should be able to
generate the SDK using the MySQL Shell GUI Extension for VSCode and
directly incorporate it in their Python application.

Change-Id: I3682938ba2fd711d068d1c553d9ded5d477e960a