BUG#36508205 Fix startup issue with the backend server
The MySQL Shell GUI backend server plugin is using a deprecated API to
create a server-side SSL socket to which the clients (standalone GUI and
VS Code extension) can connect.

This API has been removed in Python 3.7, and consequently, in the
lastest Python version bundled with the MySQL Shell, which prevents the
backend server from starting, and does not allow any client to connect.

This patch introduces the changes required to accommodate the proper API
for creating a server-side SSL Socket in the lastet Python versions.

Change-Id: I1a2a01e5488c52ba4e84b693bdd9bcf9571c4c1b