BUG#36590424 Generate "*Cursors" type placeholders
If a database object such as a table, part of a REST service, does not
contain any column eligible to be used as the reference field for
cursor-based pagination (i.e. is not unique or not sequential), the
resulting SDK files will contain references to an undefined "*Cursors"
type declaration for the corresponding database object in different API
commands.

In order to address this issue and to avoid introducing intricate
conditional logic in the base TypeScript SDK template file, this patch
ensures that a "*Cursors" type declaration is created for each database
object under the circumnstances mentioned above. The type declaration
needs to use something that represents the type of values that can never
occur, which in the case of TypeScript, can be done with the  "never"
keyword.

Change-Id: I2893a846c0594ccb7e2d3420bc25cf34dc4d9288