Implemented response handling for non-active execution contexts
Before this patch, when switching away from a connection tab with running requests, the results were ignored and hence not available when switching back to that connection tab. Now all contexts stay alive and are taking results, even if their owning connection tab (and so the code editor) have been unmounted. For less memory consumption row data is removed from the contexts and restored on activation of the contexts (from the app DB).

Change-Id: I976d3d799e8de877bf1eab2d3819b62a14a12bbf