BUG#36925323 fix type check error in MRS TS SDK suite
The generic IDeleteOptions type definition in the MRS TypeScript SDK
currently requires two arguments. One of those arguments should be
optional, by nature. There are some usage instances in the SDK test
suite that only use one argument leading to a corresponding type check
error.

This patch updates the IDeleteOptions type definition to formally allow
an optional second argument and still be used as is in the test suite.

Change-Id: I0b2db10410eaacf52048f3452644a4875ced7368