WL#16481 Add support for MRS auth in the Python SDK
If a REST object requires authentication, an application must be
authenticated to be able to call MRS SDK commands on top of that object.

This patch introduces the "authenticate" command on the service level
for the Python MRS SDK, which an application must use when the service
contains objects that require authentication. For now, this command only
supports the native MRS authentication mechanism.

This command is only available if one or more objects in the service
requires authentication. Additionally, the command only allows to use
authentication apps that use the native MRS authentication mechanism.

Change-Id: I69d05a59948a0719b5aba1850f47c123604d742f