Fix some e2e tests on Windows
Sometimes the test "Copy REST Object Request Path to Clipboard" fails because the REST object path is not copied to the clipboard. It was introduced the repetition of this action for 15seconds for successful results.
Editing char datatype columns of result grids, fails to update the complete wanted value. It was changed the way the input field is set, as the reason of the failures is unknown

Change-Id: I9fd20b7c998a90a76a73ad6ab7ccbbe29f6bac4d